Finished startCapture for V4L2 devices with USERPTR
Currently we allocate a new Haskell GC-managed buffer for each frame and
store it in a mutable hash table while it is waiting in the driver
queue. We wait and receive frames in a separate thread and create a new
thread for each frame.

Theoretically it should be efficient enough. Otherwise we'd need to
benchmark it.